backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
numpy==2.1.3
opencv-python-headless==4.10.0.84
pillow==11.0.0
torch==2.5.1+cpu
torchvision==0.20.1+cpu
